refactor button actions
