+++ src/Bot.ts
@@ -46,7 +46,6 @@ export default class Bot {
     bot.on("left_chat_member", TGEvents.onUserLeftGroup);
     bot.on("chat_member", TGEvents.onChatMemberUpdate);
     bot.on("my_chat_member", TGEvents.onMyChatMemberUpdate);
-    bot.on("callback_query", TGEvents.onCallbackQuery);
 
     // action listeners
     bot.action(
@@ -57,6 +56,10 @@ export default class Bot {
       /^leave_confirmed_[0-9]+$/,
       TGActions.confirmedLeaveCommunityAction
     );
+    bot.action(/;PickRequirement$/, TGActions.pickRequirementAction);
+    bot.action(/;ListVoters$/, TGActions.listVotersAction);
+    bot.action(/;UpdateResult$/, TGActions.updateResultAction);
+    bot.action(/;Vote$/, TGActions.voteAction);
 
     // starting the bot
     bot.launch({
